+++ RxCocoa/iOS/UILabel+Rx.swift
@@ -34,6 +34,25 @@ extension UILabel {
             }
         }
     }
+
+    /**
+    Bindable sink for `attributedText` property.
+    */
+    public var rx_attributedText: AnyObserver<NSAttributedString?> {
+        return AnyObserver { [weak self] event in
+            MainScheduler.ensureExecutingOnScheduler()
+
+            switch event {
+            case .Next(let value):
+                self?.attributedText = value
+            case .Error(let error):
+                bindingErrorToInterface(error)
+                break
+            case .Completed:
+                break
+            }
+        }
+    }
     
 }